// jobhunter-FE/src/types/backend.d.ts
    export interface IBackendRes<T> {
        error?: string | string[];
        message: string;
        statusCode: number | string;
        data?: T;
    }

    export interface IModelPaginate<T> {
        meta: {
            page: number;
            pageSize: number;
            pages: number;
            total: number;
        },
        result: T[]
    }
// ...
    export interface ICompany {
        id?: string;
        name?: string;
        address?: string;
        logo: string;
        description?: string;
        createdBy?: string;
        isDeleted?: boolean;
        deletedAt?: boolean | null;
        createdAt?: string;
        updatedAt?: string;
    }